Update dependencies in Cargo.toml + build instructions
I was able to get the blinky example to build and run on a
SAML11 Xplained board with these changes.

Looks like you were able to upstream the `thumbv8m.base-none-eabi`
target definition:

https://github.com/rust-lang/rust/pull/55041

I was able to use that via rustup on a recent nightly, so no more need
for (patched) xargo.

I went ahead and bumped all of the crates to the latest releases as I
wasn't able to get anything to compile without doing so.

All that said, I was able to blink an LED on a SAML11 Xplained.